fix memory leaks

+++ src/adapter/driver/driver_internal.h
@@ -65,7 +65,7 @@ int neu_adapter_driver_validate_tag(neu_adapter_driver_t *driver,
                                     const char *group, neu_datatag_t *tag);
 
 int neu_adapter_driver_add_tag(neu_adapter_driver_t *driver, const char *group,
-                               neu_datatag_t *tag, int interval);
+                               neu_datatag_t *tag, uint16_t interval);
 int neu_adapter_driver_del_tag(neu_adapter_driver_t *driver, const char *group,
                                const char *tag);
 int neu_adapter_driver_update_tag(neu_adapter_driver_t *driver,
